+++ src/football/WRPlayer.java
@@ -12,8 +12,8 @@ public class WRPlayer extends Player {
 	public static final double POINTS_WITHOUT_TDS_LAST_WR = LAST_WR_DRAFTED_BASELINE
 			- LAST_BASELINE_TDS * 6;
 
-	public WRPlayer(String name, int dollarValue, int rushYards, int rushTDs,
-			int recYards, int recTDs, int turnovers) {
+	public WRPlayer(String name, int dollarValue, double rushYards, double rushTDs,
+			double recYards, double recTDs, double turnovers) {
 		super(name, dollarValue, 0, 0, rushYards, rushTDs, recYards, recTDs,
 				turnovers);
 		pointsRelativeToBaseline = (this.totalPoints - BASELINE_WR_SCORE);
